Update Jest tests
